Move the text-oriented helpers to the appropriate module.
